feat: add generic trigger manager and hooks
